+++ NEWS.md
@@ -1,4 +1,4 @@
-# arkhe 1.1.0.9000
+# arkhe 1.2.0
 ## New classes and methods
 * Add `math_lcm()` and `math_gcd()` to compute the lowest common multiple and the greatest common divisor.
 * Add `interval_hdr()` and `interval_credible()` to compute the credible intervals.